change workload file structure
